server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
google-generativeai>=0.8.0
